Fixed shader format to write the shader code length as 32bit integer. Incremented MGFX version to 5.

+++ Tools/2MGFX/DXShaderData.writer.cs
@@ -9,7 +9,7 @@ namespace Microsoft.Xna.Framework.Graphics
         {
             writer.Write(IsVertexShader);
 
-            writer.Write((ushort)ShaderCode.Length);
+            writer.Write(ShaderCode.Length);
             writer.Write(ShaderCode);
 
             writer.Write((byte)_samplers.Length);
